[FP8] Test e4m3 x4 vector quant/dequant * Add TestFP8e4x4QuantDequantScale(BaseFP8E4M3QuantScaleOnly) * Move fp8 tests to codegen tests
